Mock IDB handler to fix test
