+++ web/src/js/authentication/__mocks__/firebaseIDBErrorManager.js
@@ -1,0 +1,3 @@
+/* eslint-env jest */
+
+export const checkForFirebaseIDBError = jest.fn(() => Promise.resolve(null))
